backend/app/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Database Migrations
alembic==1.12.1

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI & LLM
openai
langchain-openai==0.0.2

# Vector Database
chromadb==0.4.18

# File Processing
aiofiles==23.2.1
pypdf==3.17.1
python-docx==1.1.0

# Utilities
python-multipart==0.0.6

